✨ refactor: Update import paths for QueryParams to use new query.types file
