+++ src/modules/recipeBook/services/category.service.ts
@@ -1,6 +1,6 @@
 import { PrismaClient } from "@prisma/client";
 import { Context } from "../../../shared/jest/context";
-import { QueryParams } from "../../../shared/prisma/interfaces/Iquery";
+import { QueryParams } from "../../../shared/prisma/interfaces/query.types";
 import { getQueryOptions } from "../../../shared/prisma/utils/prisma.utils";
 import { CategoryModel, CategoryCountModel } from "../models/category.model";
