+++ src/shared/prisma/interfaces/query.types.ts
